created updateUser in user service,integrated authController and authService, deleted OTPController and merged functions into AuthController
