--- .env
@@ -1,11 +1,0 @@
-PLATFORM=Express
-# Database configuration
-DB_USER=sa
-DB_PASSWORD=EZPZOSAdmin!
-DB_SERVER=localhost
-DB_NAME=EZPZOS
-DB_TRUST_SERVER_CERTIFICATE=true
-LOG_LEVEL=6
-
-# Application port
-PORT=8000
