+++ .gitignore
@@ -6,3 +6,4 @@ package-lock.json
 /Log
 package-lock.json
 package-lock.json
+.env
