+++ src/routes/routes.ts
@@ -1,6 +1,5 @@
 import { Router } from "express";
-import { sendOtp, verifyOtp } from "../controllers/OTPController";
-import { mobileLogin, signup } from "../controllers/AuthController";
+import { mobileLogin, signup, sendOtp, verifyOtp } from "../controllers/AuthController";
 import AuthMiddleware from "../middlewares/AuthMiddleware";
 import BookingRouter from "./private/Booking";
 import UserRouter from "./private/User";
