+++ src/services/UserService.ts
@@ -1,7 +1,6 @@
 import { UserRepository, RoleRepository, RoleCode, User, UserRole, Role } from "ezpzos.core";
 
 export class UserService {
-
 	// Utility method to get user repository
 	static async getUserRepository() {
 		const userRepositoryType = (await UserRepository())?.UserRepository;
@@ -20,20 +19,38 @@ export class UserService {
 		return new roleRepositoryType();
 	}
 
-	// Create and save a new user into database
+	// Create and save a new user into the database, and assign a role before saving
 	static async createUser(
 		userData: Partial<User>
 	): Promise<{ user: User | null; result: boolean; errorCode?: number; errorMessage?: string }> {
 		try {
 			// Create the user object with the provided data
 			const userRepo = await this.getUserRepository();
+			const roleRepo = await this.getRoleRepository();
+
 			const user = new User();
 			Object.assign(user, userData);
 			user.Salt = "";
 			user.IsDeleted = false;
 			user.Avatar = "[binary,...,..]";
 
-			// Save the user object to the database
+			// Assign a user role before saving the user
+			const role = await roleRepo.GetRoleByCodePromise(RoleCode.User.toString()); 
+			if (!role) {
+				return { user: null, result: false, errorCode: 500, errorMessage: "User role not found" };
+			}
+
+			// Create a new UserRole and associate it with the user
+			const userRole = new UserRole();
+			userRole.RoleId = role.Id;
+			userRole.Role = role;
+			userRole.UserId = user.Id; // Assign the userId to the userRole
+			userRole.IsDeleted = false;
+
+			// Assign the UserRole to the user object before saving
+			user.UserRoles = [userRole]; // Assign roles to user before saving
+
+			// Save the user object along with the UserRole to the database
 			return new Promise(resolve => {
 				userRepo.Save(
 					user,
@@ -63,87 +80,67 @@ export class UserService {
 		}
 	}
 
-	// assign UserRole to a new user
-	static async assignUserRole(user: User): Promise<UserRole | null> {
+	//update existed user in the database
+	static async updateUser(
+		userData: Partial<User>
+	): Promise<{ user: User | null; result: boolean; errorCode?: number; errorMessage?: string }> {
 		try {
-			const roleRepo = await this.getRoleRepository();
-			let role = await roleRepo.GetRoleByCodePromise(RoleCode.User.toString());
+			const userRepo = await this.getUserRepository();
 
-			let userRole = new UserRole();
-			userRole.Role = role ?? new Role();
-			userRole.UserId = user.Id;
-			userRole.RoleId = userRole.Role.Id;
-			userRole.IsDeleted = false;
+			// Retrieve the existing user from the database
+			if (!userData.Id) {
+				throw new Error("User ID is required for updating");
+			}
 
-			user.UserRoles = [userRole];
-			return userRole;
+			const existingUser = await new Promise<User | null>((resolve, reject) => {
+				userRepo.GetUserById(userData.Id!, (result: boolean, user: User | null | undefined) => {
+					if (result && user) {
+						resolve(user);
+					} else {
+						resolve(null);
+					}
+				});
+			});
+
+			if (!existingUser) {
+				return {
+					user: null,
+					result: false,
+					errorCode: 404,
+					errorMessage: "User not found"
+				};
+			}
+
+			// Merge the updated fields into the existing user
+			Object.assign(existingUser, userData);
+
+			// Save the updated user object to the database
+			return new Promise(resolve => {
+				userRepo.Save(
+					existingUser,
+					existingUser.Id,
+					true, // Indicating an update operation
+					false,
+					(result: boolean, errorCode?: number, errorMessage?: string) => {
+						if (result) {
+							// If the user is successfully saved, resolve with the updated user object and result
+							resolve({ user: existingUser, result, errorCode: undefined, errorMessage: undefined });
+						} else {
+							// If there was an error saving the user, resolve with error details
+							resolve({ user: null, result: false, errorCode, errorMessage });
+						}
+					}
+				);
+			});
 		} catch (error) {
-			console.error(error);
-			return null;
+			// Handle any exceptions that occur during the user update process
+			console.error(`Error during updateUser: ${error}`);
+			return {
+				user: null,
+				result: false,
+				errorCode: 500,
+				errorMessage: "An unexpected error occurred during user update"
+			};
 		}
 	}
-
-  //update existed user in the database
-	static async updateUser(
-    userData: Partial<User>
-  ): Promise<{ user: User | null; result: boolean; errorCode?: number; errorMessage?: string }> {
-    try {
-      const userRepo = await this.getUserRepository();
-  
-      // Retrieve the existing user from the database
-      if (!userData.Id) {
-        throw new Error("User ID is required for updating");
-      }
-  
-      const existingUser = await new Promise<User | null>((resolve, reject) => {
-        userRepo.GetUserById(userData.Id!, (result: boolean, user: User | null | undefined) => {
-          if (result && user) {
-            resolve(user);
-          } else {
-            resolve(null);
-          }
-        });
-      });
-  
-      if (!existingUser) {
-        return {
-          user: null,
-          result: false,
-          errorCode: 404,
-          errorMessage: "User not found"
-        };
-      }
-  
-      // Merge the updated fields into the existing user
-      Object.assign(existingUser, userData);
-  
-      // Save the updated user object to the database
-      return new Promise((resolve) => {
-        userRepo.Save(
-          existingUser,
-          existingUser.Id,
-          true, // Indicating an update operation
-          false,
-          (result: boolean, errorCode?: number, errorMessage?: string) => {
-            if (result) {
-              // If the user is successfully saved, resolve with the updated user object and result
-              resolve({ user: existingUser, result, errorCode: undefined, errorMessage: undefined });
-            } else {
-              // If there was an error saving the user, resolve with error details
-              resolve({ user: null, result: false, errorCode, errorMessage });
-            }
-          }
-        );
-      });
-    } catch (error) {
-      // Handle any exceptions that occur during the user update process
-      console.error(`Error during updateUser: ${error}`);
-      return {
-        user: null,
-        result: false,
-        errorCode: 500,
-        errorMessage: "An unexpected error occurred during user update"
-      };
-    }
-  }
 }
